Extracting an interface for IngestionRequest handling
- Reuse the IngestionService code for the DHT nodes and the proxy
